requirements.txt
aliyun-python-sdk-core-v3==2.11.0
amqp==2.3.2
aniso8601==4.0.1
billiard==3.5.0.4
blinker==1.4
celery==4.2.1
Click==7.0
Flask==1.0.2
Flask-Limiter==1.0.1
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.2.1
limits==1.3
MarkupSafe==1.1.0
pycryptodome==3.7.0
PyJWT==1.6.4
pytz==2018.7
redis==3.0.1
six==1.11.0
SQLAlchemy==1.2.14
vine==1.1.4
Werkzeug==0.14.1
